Use `typing.ClassMethod` to remove special casing for `@classmethod`
Summary: Now that we have a class definition with the real descriptor behavior encoded on it, we can now exploit it to treat `classmethod`s as attributes, not special cased as `Method`s.  With all the descriptor infra, this is as simple as having `create_attribute` emit a `typing.ClassMethod[callable_type]` attribute instead of a `callable_type` method.

Reviewed By: grievejia

Differential Revision: D21186776

fbshipit-source-id: 0d4cc5d589ac35f341b5e581bc88c88a48e11cae