Skip marking escapees when resolving __get__
Summary:
Escaped free type variable marking is an element of the long-dormant plan to error on incomplete types being assigned to names.  This behavior is becoming problematic in the context of descriptors such as `classmethod` and `staticmethod` and ultimately `Callable` itself.  The issue comes from situations as simple as this one:

```
C = TypeVar("C")
T = TypeVar("T")
class staticmethod(Generic[C]):
  def __call__(self, f: C) -> None: ...
  def __get__(self, host: object, host_type: object = None) -> C: ...

class F:
  staticmethod
  def generic(x: T) -> T:
    return T
```

When we'd resolve the `__get__` in this instance, we'd get a return type of `Callable[[T(free)], T(free)]`.  These free variables would then be marked as escaped, and potentially be collapsed out of unions.  The real fix to this kind of situation is some sort of active variable scoping, or my long-term plan of abstracting function types behind function names.  Until either of those come to pass, it's just easier to suppress this behavior in this case.

Reviewed By: grievejia

Differential Revision: D21211368

fbshipit-source-id: d9550be92b6987b177369603cb069c6c88141ae0