Refine `is not None` assertion for union types
Summary:
We used to support refining the type of `x` in `if x is not None` only when `x` has an optional type. The refinement does not work when `x` has a union type that contains `NoneType`. Let's fix that.

NOTE: Prior to D21163447, since we allow non-canonical types of the form `Optional[Union[X, Y]]`, it may appear that we sometimes do support refining unions. But this kind of "support" is super fragile and non-generalizable.

Reviewed By: shannonzhu

Differential Revision: D21240409

fbshipit-source-id: 774361c143e810152ef22814880dd81b2db011d7