Preserve expression location when processing `assert x is not None`
Summary:
When seeing this statement:
```
if x is None: ...
```
We'll transform it into the following in CFG construction:
```
if (*):
  assert x is None
  ...
else:
  assert x is not None
  ...
```
In the else branch, we use a synthetic location for the fabricated condition `x is not None` to prevent any type error from being surfaced there (see D20684546).

However, when processing `assert x is not None`, we'll just transform it into `assert x`. This is problematic for many reasons, but the one reason that's relevant here is that the location for `x` is not marked as synthetic as the location for `x is not None` -- meaning that the suppression mechanism introduced in D20684546 would fail here!

This diff simply makes it such that we preserve the synthetic location of `x is not None` when converting it to `x`. In principle, we shouldn't do the conversion in the first place --but I'll leave that as future work.

Reviewed By: dkgi

Differential Revision: D21318463

fbshipit-source-id: 374f250b796af4696859d6b1a3d14ad7ecb6acfb