special case `x = None` assignments for always- breadcrumbs
Summary: The problem: We know that when `x = None`, there is no forward taint associated with a variable. If there are some branches in the code which lead to a source of taint, the source breadcrumbs are `via-` instead of `always-via` due to the presence of the None. By simply clearing existing taint when we see `x = None`, we can ensure that if any other branches have taint, they get the always- annotation, and that we properly clear taint when None is assigned still.

Reviewed By: dkgi

Differential Revision: D21361972

fbshipit-source-id: e31fcc830e438589844ae8a3ce2137e8156d5429