Fix stdin vs generate source option for upgrade errors
Summary:
catsigh

Turns out I didn't really understand argparse boolean values -- it's impossible to pass false into a flag, makes sense that we want these to be additive but does make it tough to have a standardized `--from-stdin` flag across different commands when the default for one is true and the default for another is false.

Walking back that change by letting people pass in a `--error-source stdin` or `--error-source generate` if they want to change the default behavior of a particular command. Another option is just to revert the other diff and use `--from-stdin` for one command and `--run` for the other. :P

Reviewed By: MaggieMoss

Differential Revision: D21383047

fbshipit-source-id: 546a3cee0fef6af39442af7d47d0eeeac28c3b49