Callable('name', T) </: Callable('other_name', T)
Summary:
Context: If two named callables have the same signature, we currently join them to be a single callable, but only keep one of the names around. This means that Pysa doesn't have enough type info to simulate both calls. In practice, this usually doesn't matter (unless we're dealing with locals getting assigned two different methods/functions based on an if statement),  so let's keep the extra information around.

Longer term, we'd like to move towards having the name of a callable be the entire type (which you lazily dereference in signature select), and this moves our behavior towards that.

Also adds tests demonstrating how this helps Pysa catch new flows, and how the callgraph now detects both possibilities.

Reviewed By: dkgi

Differential Revision: D21407220

fbshipit-source-id: 98f3f344c2f15e6820db13f53d3ff928a5c8739e