Add required no-op `register` function for dependency keys
Summary:
As described in D21362097, it will be nice to have the support of the OCaml type system to ensure that we're registering all of the keys that we ever need to decode.  For now, there is no such registration process and thus `register` is a no-op.  However, in this diff we ensure that every key we use has been passed through this no-op by way of a sealed `registered` type.

Accordingly this diff just involves plumbing these "registered" dependencies around.

Reviewed By: grievejia

Differential Revision: D21362551

fbshipit-source-id: 29829055ef7364f0c9dff5624318686b44bf3315