Factoring out decode logic into each dependency key
Summary:
The core idea of this stack is that we can save some saved state space, heap size and even get some more speed by being more creative about how we do encoding and decoding.

Here's a bit of background on the status quo:
* In order to store our dependencies in shared memory, we use two things: normal shared memory tables and the dependency table.
* The shared memory tables are write-only in the workers, but can have any kind of keys/values
* The dependency table is appendable in workers but can only associate 32 bit integers.
* When we're querying the dependency table, we have to convert to and from these integers from the full space of keys we might use.
* We currently do this in a 1-to-1 fashion via the encode and decode tables.  The encode table is an (dependency key -> int) table which assigns what is effectively a UUID to each key, and the decode table is just the inverse of that.
* When we record a dependency from k1 to k2 (i.e. k1 depends on k2), we add hash(k1), encode_table(k2) to the dependency table.
* If k1 is triggered, we can query the table for hash(k1) and then return decode_table(encode_table(k2)) = k2 (along with other keys).

This diff doesn't actually change any of that but instead just opens up the space for such experimentation.  An example of what that experimentation could look like is demonstrated in `dependencyTrackedMemory.ml`, where instead of an encode and decode table in shared memory, we instead have just a decode table, this time stored as a global.

The key insight here is that if we take this out of shared memory, we don't need an encode table, as we can enforce that encode(x) = hash(x).  This introduces the chance for collisions, which we can now deal with in a 1-to-many data structure in non-shared memory.  I'll attempt to apply this insight to our main dependency key in the following diff.

Another important decision here is the introduction of the `registered` type.  Moving models like this will involve minimizing the number of times we re-add these entries to our decode data structures, but still ensure we're doing so when we need to.  This is best supported by some OCaml typing.
(Down the road we may be able to remove this but for now its very helpful)

Reviewed By: grievejia

Differential Revision: D21362097

fbshipit-source-id: 359b77c470c4014e2a71ca0384e44cb1c75e9d6a