Include module in unannotated global assign location
Summary:
> The way the current environment stores locations means we only keep track of in-file locations (ie., line and column information), and tack on paths to the locations during type checking where we are looking through a particular source with a qualifier.
We store a global location table with location information of globals, but there is no good way, given a fully qualified global name, to include the path in the location.
This is problematic for both location queries as well as go to definition, which is currently disabled but would not be working if we turned this on.
With a fully qualified name we can't tell what part of the qualifier is the path to the module which should go into the location, and what part of the qualifier is, say, a parent class. a.b.c.Foo could mean c.Foo inside a.b.py or Foo inside a.b.c.py.
Adding qualifier information to class summaries and unannotated globals so that we can have a full location (Location.WithModule.t) available in global location lookups rather than just line/column information.

Reviewed By: dkgi

Differential Revision: D21295378

fbshipit-source-id: 22d09e445a1cf0940c8fadc9430810f6ff9f4463