More precise locations for names bound with imports
Summary:
When I started writing the scope analysis, we did not keep track of the locations of each individual import alias. So I was forced to use the location of the entire import statement for each introduced bindings:

```
from foo import bar, b as baz

# Binding of `bar` and `baz` points to the location of the import statement.
```

But after D19038151 we started tracking locations for import aliases. Let's utilize the to give us more precise location info.

This change doesn't really affect any user-visible behavior at the moment. I did it just to get rid of a TODO comment in the code :P

Reviewed By: pradeep90

Differential Revision: D21513075

fbshipit-source-id: db6ac0a20c70c32161587322a384bba7a4163231