Reduce reachable trace frames by leaf kinds
Summary:
When generating the traces for an issue we can stop potentially earlier, when the trace leads to frames that no longer have the kinds of leaves we are looking for.

Note that we still produce some frames at the boundary of reachable that may not be used (in order to determine if they need to be traversed).

Reviewed By: zlandau

Differential Revision: D21487175

fbshipit-source-id: 02056e5e3cd36e708fdd224cdc8d42d0bd5b9fa2