Unbreak targets to configuration by adding arguments
Summary:
Command was broken because the parent initializer expects a `max_line_length` argument that is set in the parent arguments parser but not in this child.

Might be a result of moving args into fixme and calling the parent argument parser there; should do the same for other children of ErrorSuppressionCommand to fix them.

Reviewed By: dkgi

Differential Revision: D21539912

fbshipit-source-id: b5a91fa6ae6ca6316ce3fbb268d464c2dcb92df5