Unwrap BoundMethods when comparing against protocol methods
Summary:
We previously added special case handling to instantiating a method that would "unwrap" (i.e. replace it with the signature of its `__call__` method) the `BoundMethod` types of protocol methods (D20369134).  However, as we started removing things from that codepath, it became clear that that wasn't a broad enough solution.  Because we started running things like `ClassMethod` and `StaticMethod` through as attributes instead of methods, we weren't hitting this special casing anymore.

A better approach is to do this unwrapping in the contexts where it's actually needed:
* In instantiate_protocol_parameters in order to allow methods with different names and self types to be less than one another.
* In call graph building as a performance optimization

Reviewed By: sinancepel

Differential Revision: D21511534

fbshipit-source-id: 4137af4b5a5a8153717efe7291b94c7fdda7c168