Allowing non-data-descriptors to be set according to their get type
Summary:
I initially thought that I was going to need to ban assignments on instances for non-data-descriptors.
The situation I was seeing was the following:

```
      class Descriptor:
        def __get__(self, o: object, t: object = None) -> int:
          return 1

      class Host:
        a = Descriptor()

      def f() -> None:
        h = Host()
        expects_int(h.a) # works because a is a descriptor for int
        h.a = Descriptor() # because its a non-data descriptor it's "set type" is Descriptor
        expects_int(h.a) # but now h.a will return Descriptor, not int, since items on the instance dictionary aren't "descriptored"
```

However, instead of just banning these assignments, we can just alter the set type to match the get type.  This means that the get type is always reliable.
```
      def f() -> None:
        h = Host()
        expects_int(h.a) # works because a is a descriptor for int
        h.a = 1 # we now expect an int here, not a Descriptor
        expects_int(h.a) # h.a returns an int from the instance dictionary, not from Descriptor.__get__, but that still fulfills the contract.
```

This makes life a bit more flexible, and allows us to not have to alter this existing behavior in methods as we gradually delete that type.

Reviewed By: grievejia

Differential Revision: D21529686

fbshipit-source-id: a37837ba05473599f327521539e6a95a6f0092df