Correctly handle what names `import` statement would bind
Summary:
Previously, we did not handle imports correctly when create binding for it. For example,
```
from a import b  # statement 1
import c.d  # statement 2
```

We correctly identify that statement 1 binds the name `b`. But we incorrectly identify that statement 2 binds the name `c.d` (the correct answer should be `c` only).

Let's fix that.

Reviewed By: sinancepel

Differential Revision: D21539725

fbshipit-source-id: aaf2e5b17d01866aaacf2fe57c58ee9c78879456