Include locations in collected name accesses
Summary:
The `collect_access` function, as the name suggests, collect all accessed names for the given statements for scope-related analyses. For example,
```
def foo():
  x = y + 1
  z = bar.baz(x)
```

The accessed name for `foo` in this case would be `y`, `bar`, and `x` (roughly speaking, every names appeared on the RHS of assignments).

Previously the function only collect the accessed names but not their corresponding locations. This is OK for captured variable analysis because it doesn't use the location info. But for unbound name analysis we'll want those location info in order to emit errors at the right place.

Let's include the locations when collecting accesses. This diff should have no impact on existing behaviors but it paves the way for D21532591

Reviewed By: sinancepel

Differential Revision: D21532598

fbshipit-source-id: ec71b094c27119e4d303fb3b5dc13f5fb9228e61