Judging overload compatibility on their undecorated signatures
Summary:
In the course of implementing general decorators we need to nail down the semantics of what decorating overloaded functions means.   I propose the following sematics:
* a function object corresponds to a single implementation `def` and an arbitrary number of overload `defs`, *not* including any of the decorators
* this function object is typed via parameters and return types on those `defs`.  These combine to form one `typing.Callable` type.
* if this is decorated, the entire function object is passed sequentially through the chain of decorators.   The final type of that chain is the type inferred for the attribute/global.

These semantics avoid any need to have to apply overload combination logic to the return type of a decorator, which is the fundamental blocker right now for general decorators.

With this being the case, it makes sense that we should be doing overload validation things on the undecorated overloads, not the decorated overloads.

Reviewed By: dkgi

Differential Revision: D21466734

fbshipit-source-id: 4715e56a2c0b2171e566328efdab0ecce681dec8