Storing undecorated signatures in globals and attributes
Summary:
We rely on the ability to "find" the overload currently being checked within the overall callable for things like overload compatibility.  However, we currently do so by searching for the decorated overload within the decorated signature. This will not work for general decorators, and doesn't even work right for standard decoration.

Instead we should be searching for the undecorated overload within the undecorated signature.  This requires saving that undecorated signature, which we haven't needed before.  This diff does just that, in anticipation of future diffs that will actually use that.

The impact on memory could be further mitigated by not storing both types if annotation = Callable undecorated_signature, but I'd rather work on that in a separate work item that also includes original_annotation (T66904528).

Reviewed By: dkgi

Differential Revision: D21466426

fbshipit-source-id: 397132c35801e3f4ed8c348b32abb11124097baa