Only mark local free variables as escaped
Summary:
We have a behavior in signature selection that marks free variables in the returned type as being "escaped", meaning that they have escaped the context they were meant to be contained in.

However, we were also marking free variables that were just passing through (as in a generic function parameter) as also being escaped, even though they are being returned to the context they came from.

This is wrong, and leads to strange behavior when we start using signature select for decorating generic functions.

Reviewed By: sinancepel

Differential Revision: D21554124

fbshipit-source-id: a5f770a9422c1332573699657344d799b809f65d