Only propagate a decorated callable's name when the type exactly matches
Summary: D21486250 preserved a problematic existing behavior that would always blindly apply the original name of the callable onto the decorated callable.  This isn't safe when we rely on the name implying a certain type signature.  Its not exactly great even when it is the exact same type, but permitting it in that particular case is a reasonable compromise between taint analysis and type checking.

Reviewed By: dkgi

Differential Revision: D21487479

fbshipit-source-id: 3b21a2c9ba2816c4f2c737bfbb066b876136e4fd