apply decorators to the combined callable
Summary:
As described in D21466734, we'd like to implement decorator semantics where we apply the decorator chain to the combined overloaded callable all together.  This diff does this by parsing the overloads, combining them into a callable type, then applying the decorators once, rather than once per each overload.

This results in one return type, rather than a set we have to combine.  This is the last hard blocker for general decorators.

Reviewed By: sinancepel

Differential Revision: D21486250

fbshipit-source-id: 7639edae58dc45f694a29776393f65c74d63f746