DifferingDecorators error
Summary:
With the decorated overloaded function semantics defined in D21466734, we're required to have the same set of decorators for all of the overloads.  However, enforcing this is pretty tricky.
Naively we'd like to just search for the other defines, but doing so, especially for methods, is pretty tricky.  On the other hand, there is a site where we explicitly have access to this information: in the actual environment building.  Routing between these two sites requires doing something we've long wanted to do: erroring from environment building.

Erroring with `AnalysisError`s from environment building isn't actually possible, but it is possible to achieve the effect.  The trick here is to have resolve_define return a `Result`, and then have annotatedglobalenvironment and attributeresolution tag that problem, if it exists, onto the attribute or the global.  This then can be retrieved during type checking.

We'd really like to avoid putting locations into our attributes, so instead we can use the undecorated signature to implicitly find when we are working on the offending overload.

Reviewed By: dkgi

Differential Revision: D21480989

fbshipit-source-id: e35221a82d18e4d3f502bd04dbe4af470ef36802