no longer rely on name of target vs. value when determining types of class attributes in constructors
Summary: We were requiring target name = value name historically when determining whether to infer types for attributes in constructors. There's no reason to do this. The heuristic was constrained in D13928554, but the other part of the diff (do not infer types in the constructor if we're in an if statement) is sufficient to get rid of the unsoundness.

Reviewed By: dkgi

Differential Revision: D21578037

fbshipit-source-id: 3ba7ed95af4308008ed7c2cd37cbb499f133ec99