Show paths where upgrade error suppression failed
Summary:
Ran into this error when suppressing errors in a project, and wished I knew which files the failures occurred in.

It looks like the way this custom exception is structured is setting paths as an attribute of the exception but then also taking in a message into the constructor rather than generating one with a custom `__str__`.... going to assume the paths attribute is there bc it'll be used for something later, and will just change the message being passed in.

Reviewed By: grievejia

Differential Revision: D21560046

fbshipit-source-id: 4358e60a79116dfde80e5c17039d1608edeefff9