Treating staticmethod as a normal decorator
Summary:
Now that we support general decorators, we can start taking advantage of this to start folding some special cases closer to being handled natively.

Next up is staticmethod, which was previously being handled uniquely by detecting the presence of the staticmethod decorator outside of the normal chain logic and wrapping the final result.  This only works correctly if staticmethod is the last decorator in the chain.  This diff moves handling of the decorator into the normal chain flow, and handles the magic `__new__` correctly

NOTE: we can't currently just write this as a stub, because staticmethod is actually a non-generic class.  We'll have to modify typeshed/actual typing, as well as support class construction as decoration in order to plain native support this.

Reviewed By: pradeep90

Differential Revision: D21577864

fbshipit-source-id: a4b94c29f8603e5812c0c81fd70dc9f0188e6f34