Handle classmethod as a decorator
Summary:
Now that we support general decorators, we can start taking advantage of this to start folding some special cases closer to being handled natively.

The first of these is `classmethod`, which was previously being handled uniquely by detecting the presence of the classmethod decorator outside of the normal chain logic and wrapping the final result.  This only works correctly if `classmethod` is the *last* decorator in the chain.  This diff moves handling of the decorator into the normal chain flow, and handles the magic "__init_subclass__" and "__class_getitem__" correctly

NOTE: we can't currently just write this as a stub, because `classmethod` is actually a non-generic class.  We'll have to modify typeshed/actual typing, as well as support class construction as decoration in order to plain native support this.

Reviewed By: pradeep90

Differential Revision: D21576141

fbshipit-source-id: 97334284efa0c93d2190d7e2dc1c956bb303cb30