Validate { **this_should_be_a_mapping }
Summary:
We currently enforce that double-starred arguments to functions are Mapping[str, X]s, but haven't been doing so for the double starred arguments that you can put into a dictionary literal.

This diff repurposes that pre-existing error code for this slightly different purpose (necessitating a flag to slightly alter the message).

Reviewed By: dkgi

Differential Revision: D21579106

fbshipit-source-id: 94104a2f5d4738f0b7feaf58209f016b4041f709