Respect the return type of __new__ when it is final.
Summary:
Problem: We want to support `__new__` overloads. Currently, we always set the return type of `__new__` and `__init__` as `Foo[T]`. This means that if `__new__` or one of its overloads returns `Foo[Optional[T]]`, we won't use it.

In order to be minimally invasive, we respect the return type of `__new__` only when the class is Final. Otherwise, we fall back to the old behavior of making the return type `Foo[T]`.

Note that the old behavior was to keep the overloads around. But since the return type was always made `Foo[T]`, the overloads served only to eliminate invalid argument types.

Reviewed By: mrkmndz

Differential Revision: D21478768

fbshipit-source-id: 75a429073fb325336334295f1b89670a37b0b9f6