Don't consider changed search path files as new files.
Summary: Problem: To decide whether a file was new, we checked only if an updated file was absent from the link tree. However, since files from the search path are not necessarily in the link tree and not necessarily tracked by Buck targets, this meant that changing a file in the search path would mark it as a new file and then, after a buck query, ignore it as untracked.

Reviewed By: dkgi

Differential Revision: D21565318

fbshipit-source-id: edef46d81dc101249d03d93ae9d0d8aa58eb70e8