Determine 'is_test' on each class
Summary:
We support special test constructor functions like `setUp` adding attributes to a test class.  Currently, mimicking the runtime, when a class inherits from a test base class, we register attributes in its setUp method, as well as in all of its parents **even when those parents are not tests themselves**.  This is kind of weird behavior since that class will have errors in it since we aren't recognizing its setUp as being legitimate, but we are recognizing it in its child.  This also means we need to key all attribute tables on whether they're being accessed through a child class that's a test, which makes the key bigger, and decreases the cache hit rate.

This diff changes our semantics to only pick up `setUp` in a class that is itself a test, producing consistent behavior between children and parents.

Reviewed By: grievejia

Differential Revision: D21560767

fbshipit-source-id: 064cbd4f66adb73ca20d17e101f69debf5a639f5