Make RefinementUnit `meet` include any extra attributes.
Summary:
Problem: RefinementUnit's `meet` on `foo.x` and `foo.x.y` would ignore the `y` and just return `foo.x`. This meant that we couldn't refine stuff like `if not foo.x or not foo.x.y: ...`

Solution: When folding the two attribute trees, if one is empty, include the other tree.

Reviewed By: dkgi

Differential Revision: D21595377

fbshipit-source-id: 16fd686888cdedeeb23cf7000a61cf3f802cb364