Only report unbound name once for each name
Summary:
This is a somewhat debatable decision: if the same unbound name gets referenced multiple times, do we want to error multiple times? For example:
```
def foo(x):
  if x:
    return derp + 1
  else:
    return derp + 2
```
Do we emit one error on line 3, or do we emit 2 errors on line 3 and 5?

I can see arguments on both ways, and we also don't have a consistent preference for similar kinds of errors. So the decision is kinda arbitrary. In this diff I choose to emit error only once to conserve some space in shared memory :P

Reviewed By: dkgi

Differential Revision: D21566077

fbshipit-source-id: 41bcf0ddf7a0633f74fe2703866ac7df3ce3e686