Deduplicate UnboundName error and UndefinedType error in postprocessing
Summary:
The `UnboundName` error and `UndefinedType` error tends to overlap a lot. We often emit both the `UnboundName` and the `UndefinedType` error simultaneously for the same name, which is not ideal.

On the other hand, neither one of those two 100% covers the other one:
- The `UnboundName` error covers cases like undefined import, while the `UndefinedType` does not.
- The `UndefinedType` error covers cases like this:
```
MyType = 1
x: MyType = ...
```
while the `UnboundName` does not.

To avoid double-erroring while keeping both kinds of errors, I choose to deduplicate in the postprocessing phase. The idea is to discard the `UndefinedType` error if `UnboundName` error exists on the same name. We've already done this for `UndefinedName` and `UndefinedImport` errors so the infra is already there, which is super nice.

Reviewed By: dkgi

Differential Revision: D21566090

fbshipit-source-id: 2a070d1f61e2fc0ce270d4c6695b924e33be6e2a