Avoid duplicate error for a < b in control blocks.
Summary:
Problem: When we see an `if <foo>:` block, we add `assert <foo>` to the if-branch and `assert (not <foo>)` in the else-branch. If there's a type error in `<foo>`, then even though it is emitted twice, I believe it gets deduplicated in the error map.

Not so with `a < b` and `a >= b` (from `not (a < b)`), which have slightly different error messages due to the inverse operator.

So, ignore all errors from the assert in the else-branch.

Reviewed By: dkgi

Differential Revision: D21595509

fbshipit-source-id: bf9b38df6e25603ecb04fe5bb385e5e70b5adfab