Remove `Location.synthetic`.
Summary:
Problem: The only place this was used was in suppressing errors in else-branch, fabricated assertions. Now that we directly check for them in D21595509, we can remove `Location.synthetic`.

NOTE: The issue removed from `conditional.py.models` is a duplicate of the issue just above it, except with a synthetic location. I believe that is because the taint analyzer did not ignore the else-branch assertion and thus errored twice :P

Reviewed By: grievejia

Differential Revision: D21605778

fbshipit-source-id: e15f7869b70aee21642ff705f667779542f76d5e