require @overload to be the topmost decorator
Summary:
Once we start erroring on particular decorators being invalid, it would be nice to have an indexing scheme so that we can refer to these overloads without including expressions with their concomitant locations.
However, in order for these indexes to be reliable on overloaded functions, we need to have a consistent position for the `overload` decorator (see D21605802 for how that is useful).
I arbitrarily picked the top as that looked most natural to me.

Reviewed By: grievejia

Differential Revision: D21605016

fbshipit-source-id: b182c7915b5ad58dd7d77202a1b9491b83d6f640