Support walrus operator
Summary:
This diff introduces support for walrus operator in the scope analysis.

The main issue walrus operator brings to the picture is that expressions can now bind names. So what we need to do here:
- Add a new API `Binding.of_expression` to collect all the walrus-introduced bindings while recursing into the expression.
- Change other APIs like `Binding.of_statement` to invoke `Binding.of_expression` on all expression subtrees.
- Everything else should work as before.

I'd wish it could be as clean as that. But since we are talking about Python, it's not surprising that there are special-cases that optimizes for a subset of use cases while complicating any tools that wants to deal with the language in a more general fashion {emoji:1f63f}

To quote [PEP 572](https://www.python.org/dev/peps/pep-0572):

> An assignment expression does not introduce a new scope. In most cases the scope in which the target will be bound is self-explanatory...

> There is one special case: an assignment expression occurring in a list, set or dict comprehension or in a generator expression (below collectively referred to as "comprehensions") binds the target in the containing scope, honoring a nonlocal or global declaration for the target in that scope, if one exists.

This means that comprehensions, while previously "seals" the contained scope just like function definitions and lambdas, would now "leak" all walrus-introduced bindings into the parent scope. For example,

```
def foo():
  return [(y := x) for x in range(10)]
```
Here the name `y`, which intuitively should be a binding local to the list comprehension, must be treated as a binding in the body of `foo()` (i.e. you could access `y` in any statements that follows). This means two things for us:
- In `Binding.of_expression`, we can't blindly assume that comprehensions never introduce new bindings. Instead, we'll need to recurse into the elements and conditions looking for walrus operators as well.
- In `Scope.of_expression`, we need to filter out all walrus-induced bindings if the given expression is a comprehension, since those bindings have already been included in the containing scope.

Fortunately, the PEP authors are sane enough to introduce the following restriction:
> Due to design constraints in the reference implementation, named expressions are disallowed entirely as part of comprehension iterable expressions (the part after each "in", and before any subsequent "if" or "for" keyword)

So we are spared as we don't need to consider how this "leaking" behavior should work for comprehensions that are deeply nested...

Reviewed By: dkgi

Differential Revision: D21610879

fbshipit-source-id: 1ae4a77342d151e1348031e0abbfccecae7745f2