Use fold instead of concat for collector-like APIs
Summary:
This diff is not intended for any behavioral change -- it should be style change only.

In `scope.ml`, there are some APIs that tries to recurse on ASTs while collecting stuffs along the way. The way it used to work is that if a node has 3 children, we'll do something like this:
```
let collect { child0; child1; child2 } =
  let collected_child0 = collect child0 in
  let collected_child1 = collect child1 in
  let collected_child2 = collect child2 in
  List.concat [collected_child0; collected_child1; collected_child2]
```

This diff changes the implementation into something like this:
```
let collected sofar { child0; child1; child2 } =
  let sofar = collect sofar child0 in
  let sofar = collect sofar child1 in
  collect sofar child2
```

The new style is not only more concise, but also more efficient as it avoids traversing `collected_child0` and `collected_child1` more than once, guaranteeing a linear complexity.

One downstream consequence is that the final result is now reversed. So the consumers of the APIs need to be aware of the ordering change if it matters to them.

Reviewed By: dkgi

Differential Revision: D21610545

fbshipit-source-id: 1f1e3d9dbb0c6ca64eb609824b1541986c591ca1