Initialize NamedTuple attributes on the instance, not the class
Summary:
it turns out that the part of the NamedTuple magic is transforming assignments that are on the class to ones on the instance.  This has significance for things like descriptors.

from the runtime:
```
from typing import NamedTuple

class X:
    def __get__(*args):
        return 1

class Foo(NamedTuple):
    value: X = X()

foo = Foo()
print(foo.value) -> <__main__.X object at 0x7fc84fc39950>
```

NOTE: This means we don't infer the types for attributes with a simple default anymore, but this seems like a reasonable compromise.

Reviewed By: grievejia

Differential Revision: D21634930

fbshipit-source-id: 3706eda062e7e906a27aa59e89efd2896c7ac583