Change the type of import alias from `Reference.t` to `Identifier.t`
Summary:
I believe in the following cases:

```
import a.b as c
from a.b import c as d
```
Whatever follows the `as` keyword can never be qualified (i.e. one can't do `import a.b as c.d` or `from a.b import c as d.e`). Let's change the type of our AST to reflect that.

NOTE: Actually in `from XXX import YYY`, the `YYY` part can also never be qualified. But that would be a much more involved change I don't want to deal with for now :P

Reviewed By: dkgi

Differential Revision: D21631472

fbshipit-source-id: c76b6394fd469fb44ccd7b37803de1e7c6cf3bdf