Do not throw impossible assertion errors on not isinstance
Summary:
Remove impossible assertion errors on `not isinstance` assertions, because these are often used as a runtime sanity check to ensure errors are caught if the wrong type was passed in and there wasn't enough type coverage to catch it statically.

One drawback is Pyre cannot infer the type of the variable being asserted as anything but Bottom, and does this silently now.

Reviewed By: sinancepel

Differential Revision: D21636197

fbshipit-source-id: 5c73bd9a06610087e4901426ef10c06cb01a0796