Stop enforcing self parameter on decorated methods
Summary:
The idea of self-type parameter verification is predicated on the assumption that the method will be called "conventionally", i.e. with the first argument being called with an instance via method binding.

However, in the presence of decorators, this is not guaranteed.  The simplest way to think about this is when someone effectively makes a user-level copy of `classmethod`, without special backend support, we'll complain that they're taking a class when they should be taking an instance.

The conservative approach is to skip this check in the presence of decorators, and accept the potential unsoundness for now.

Long term we ought to be enforcing this check on the decorated type (T67529079), but for now, on balance it seems less confusing to an end user to avoid this false positive.

Reviewed By: sinancepel

Differential Revision: D21736452

fbshipit-source-id: 04259130ad1bbdfc88ce8384b122973fbcfc1262