Using real OCaml objects for our AttributeResolution open recursion
Summary:
I suspected that there was a better way to do the open recursion trick in AttributeResolution than the manual recurser record solution.
The simultaneous interest in caching metaclass lookup and adding globals into the mutual recursion prompted me to revisit how we could make this file more manageable.

To quote Real World OCaml:
> use classes and objects in situations where open recursion is a big win.

The fact that we were already using a open_recurser record meant that transitioning to a class was pretty mechanical.

The structure now is: `base` defines the uncached implementations of all of the mutually recursive functions.
we then define a managed cache that is defined in terms of the base#parse_annotation, which then can be used to make a subclass with the parse_annotation method overriden.
this process is then repeated with the uninstantiated_attribute cache.

Reviewed By: grievejia

Differential Revision: D21747890

fbshipit-source-id: aaba17783a25a0675efd1f4f4985a45c5baeea20